Refactor helper functions to typescript in core
