+++ packages/core/src/config.ts
@@ -28,7 +28,7 @@ export let config = {
   "poolFactoryAddress": "0xF9E08660223E2dbb1c0b28c82942aB6B5E38b8E5",
   "poolFactoryAddressImplementationAddress": "0x21EEC6626f15d02A8896ebB7EDD68ff3CB61e89E",
   "defaultNetworkName": "arbitrum",
-  "subgraphEndpoint": "https://subgraph.satsuma-prod.com/f37507ea64fb/xai/sentry/api",
+  "subgraphEndpoint": "https://subgraph.satsuma-prod.com/f37507ea64fb/xai/sentry/version/1.1.22/api",
   "publicRPC": "https://arb1.arbitrum.io/rpc",
   "alchemyApiKey": "oD4X3JXvJclnt36mDkqnp9CO2sZkNhYT",
 };
