+++ packages/core/src/operator/operator-runtime/operator-v1/bootOperatorRuntime.ts
@@ -23,10 +23,10 @@ export const bootOperatorRuntime_V1 = async (
         // Calculate the latest challenge we should load from the graph
         const latestClaimableChallenge = Number(openChallenge.challengeNumber) <= MAX_CHALLENGE_CLAIM_AMOUNT ? 1 : Number(openChallenge.challengeNumber) - MAX_CHALLENGE_CLAIM_AMOUNT;
         
-        const { wallets, pools, refereeConfig } = await retry(() => getSentryWalletsForOperator(operatorState.operatorAddress, { latestChallengeNumber: BigInt(latestClaimableChallenge), winningKeyCount: true, claimed: false }, operatorState.passedInOwnersAndPools));
+        const { wallets, pools, refereeConfig:refereeConfigFromGraph } = await retry(() => getSentryWalletsForOperator(operatorState.operatorAddress, { latestChallengeNumber: BigInt(latestClaimableChallenge), winningKeyCount: true, claimed: false }, operatorState.passedInOwnersAndPools));
 
         // Load all sentryKey objects including all winning and unclaimed submissions up until latestClaimableChallenge
-        const { sentryWalletMap, sentryKeysMap, nodeLicenseIds, mappedPools, refereeConfig } = await retry(() => loadOperatorKeysFromGraph_V1(operatorState.operatorAddress, wallets, pools, refereeConfig, BigInt(latestClaimableChallenge)));
+        const { sentryWalletMap, sentryKeysMap, nodeLicenseIds, mappedPools, refereeConfig } = await retry(() => loadOperatorKeysFromGraph_V1(operatorState.operatorAddress, wallets, pools, refereeConfigFromGraph, BigInt(latestClaimableChallenge)));
 
         await processNewChallenge_V1(BigInt(openChallenge.challengeNumber), openChallenge, nodeLicenseIds, sentryKeysMap, sentryWalletMap, mappedPools, refereeConfig);
         logFunction(`Processing open challenges.`);
